Delete the dangling dependency on ucspi-tcp
The inclusion of its recipe was removed several years ago in #21, but the
dependency was left behind.

If it's no longer needed, then it should be fully removed. It hasn't had a
release in quite some time, nor has the pacman cookbook it depends on that is
incompatible with Chef 14.

Signed-off-by: Jonathan Hartman <j@hartman.io>